I've mostly done everything but initial commit anyways I guess
This is a meal randomizer console program that reads into a text file that contains recipes with their urls and gives you a random meal when you ask nicely. Like I said I'm mostly done with the program except for a few things. But I decided why not make this public on git too. I'm a total beginner on cpp be nice